Added movie rate display on homepage. Added date formatter.

+++ app/src/main/java/com/rrvieira/trendyt/api/responses/PopularMoviesResponse.kt
@@ -40,7 +40,7 @@ data class PopularMoviesResponse(
         @SerializedName("video")
         val video: Boolean,
         @SerializedName("vote_average")
-        val voteAverage: Double,
+        val voteAverage: Float,
         @SerializedName("vote_count")
         val voteCount: Int
     )
